Made the by tactic more general.
Previously it could not handle λ-expressions. I fixed this by making
the apply function more general.

Until recently Agda normalised the argument to quoteTC, and this made
some invocations of the tactic (in particular, some calls to a
previous variant of the new apply function) very slow. Now apply seems
to be reasonably fast, at least for the examples I have tried.