Converted from darcs to Git.
Some tags did not depend on all previous commits. I preserved these
tags by creating separate branches for them. Note that the commits
that are only on these branches are duplicated on the master branch.
The commits on the master branch use the same (linear) ordering as the
corresponding patches used in my master darcs repository.

Darcs does not seem to record information about the local time zone in
effect when a patch is recorded. I tried to recreate this information.