Made the HIT modules more general.
These modules are now parametrised by an arbitrary definition of
"equality with J". The higher constructors of the HITs use path
equality, but many other definitions in these modules use the supplied
notion of equality.

Note that, even if the modules are instantiated with path equality,
the eliminators might now compute in a different way for higher
constructors.