Made some code type-check faster (or at all).
I did not observe large differences when I used --no-save-metas (for
this module and all its dependencies). However, when I used
--save-metas (for this module and all its dependencies in this
library), and this change was not in place, then Agda ran for a long
time before I killed the type-checker. (I had limited the amount of
heap space available to Agda, perhaps the result would have been
different with a larger limit.)